Update aerpawlib to use asyncio
It's entirely possible that there are things being missed here. Testing
is definitely a thing that's happening.

The tl;dr for future people that will wonder *why* on earth this is
being done (mainly myself) is that we needed a cleaner solution to handle
waiting for the drone to be done completing maneuvers, as well as a way
to write background tasks and hooks that aerpawlib runs.